v1.2.4.2 ========
Astropy
-------

Astropy (http://www.astropy.org/) is definitly needed, pyfits and
pywcs standalone modules are not needed anymore by ORBS (but they
still can be used by other modules ;) even modules imported by ORBS so
becarefull before removing them)

* PYFITS: now imported from astropy.io.fits
* PYWCS: now imported from astropy.wcs

Better Star fit
---------------

* :py:meth:`~cutils.multi_fit_stars`: tilted background added to the model

* detected stars are selected not too far from the center of the
  frame

* star box coeff set to 10 instead of 7 to get a better sky statistic
  around stars.

SpIOMM bias overscan for camera 2
---------------------------------

When it exists, the bias overscan created with each frame of the
camera 2 is used to remove automatically the bias. Note that in this
case **the path to the bias frames must not be given to ORBS** because
ORBS will try to create a master bias and remove it at step 3. In
fact, the mean of the master bias will be near 0 because the overscan
is removed from the bias frames also. The impact of giving the path to
the bias frame is thus not dramatic. But it is better not to give it.

Miscellaneous
-------------
* :py:meth:`~astrometry.Astrometry.register` optimization routine is
  based on a least square fit instead of a powell algorithm.

* transfered :py:meth:`~cutils.part_value` from OACS cutils.

* :py:meth:`~astrometry.Astrometry.get_alignment_vectors` simplified
  because the multi fit mode is now robust enough to remove all which
  was written for the preceding individual fit mode.

* :py:meth:`~utils.indft`, :py:meth:`~cutils.indft` added to compute
  Inverse Non-uniform Discret Fourier Transform (INDFT). New option
  **sampling_vector** in :py:meth:`~utils.transform_spectrum` to give
  the possibility to compute an INDFT by giving a non-uniform sampling
  vector.

	modifié:         doc/changelog.rst
	modifié:         doc/cutils_module.rst
	modifié:         orb/astrometry.py
	modifié:         orb/core.py
	modifié:         orb/cutils.pyx
	modifié:         orb/data/STD_MISC/README
	nouveau fichier: orb/data/filter_SPIOMM_i'.orb
	nouveau fichier: orb/data/filter_SPIOMM_i'.orb~
	nouveau fichier: orb/data/sitelle-airmass.orb
	nouveau fichier: orb/data/sitelle-airmass.orb~
	modifié:         orb/utils.py
	modifié:         orb/version.py
	modifié:         scripts/orb-reduce